requirements.txt
djangorestframework==3.12.4
requests==2.26.0
sentinelsat==1.1.1
pytz==2022.1